The createAccount method is only called after the user accepts the terms and conditions of the first step
