+++ apps/client/src/app/pages/register/register-page.html
@@ -22,7 +22,7 @@
             class="d-inline-block"
             color="primary"
             mat-flat-button
-            (click)="createAccount()"
+            (click)="openShowAccessTokenDialog()"
           >
             <ng-container i18n>Create Account</ng-container>
           </button>
