+++ apps/client/src/app/pages/register/show-access-token-dialog/show-access-token-dialog.component.ts
@@ -1,5 +1,15 @@
-import { ChangeDetectionStrategy, Component, Inject } from '@angular/core';
+import { DataService } from '@ghostfolio/client/services/data.service';
+
+import {
+  ChangeDetectionStrategy,
+  Component,
+  Inject,
+  ViewChild
+} from '@angular/core';
 import { MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { MatStepper } from '@angular/material/stepper';
+import { Subject } from 'rxjs';
+import { takeUntil } from 'rxjs/operators';
 
 @Component({
   selector: 'gf-show-access-token-dialog',
@@ -9,10 +19,16 @@ import { MAT_DIALOG_DATA } from '@angular/material/dialog';
   standalone: false
 })
 export class ShowAccessTokenDialog {
+  @ViewChild(MatStepper) stepper!: MatStepper;
   public isCreateAccountButtonDisabled = true;
   public disclaimerChecked = false;
 
-  public constructor(@Inject(MAT_DIALOG_DATA) public data: any) {}
+  private unsubscribeSubject = new Subject<void>();
+
+  public constructor(
+    @Inject(MAT_DIALOG_DATA) public data: any,
+    private dataService: DataService
+  ) {}
 
   public onChangeDislaimerChecked() {
     this.disclaimerChecked = !this.disclaimerChecked;
@@ -20,5 +36,19 @@ export class ShowAccessTokenDialog {
 
   public enableCreateAccountButton() {
     this.isCreateAccountButtonDisabled = false;
+  }
+
+  public async createAccount() {
+    this.dataService
+      .postUser()
+      .pipe(takeUntil(this.unsubscribeSubject))
+      .subscribe(({ accessToken, authToken, role }) => {
+        this.data = {
+          accessToken,
+          authToken,
+          role
+        };
+        this.stepper.next();
+      });
   }
 }
